+++ src/test/MsQuicTests.h
@@ -17,6 +17,12 @@ extern const QUIC_API_TABLE* MsQuic;
 extern HQUIC Registration;
 extern QUIC_SEC_CONFIG* SecurityConfig;
 
+//
+// Must be included under above extern declarations
+// Fixing is invasive, will do in future
+//
+#include "msquic.hpp"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
